feature: implement login mutation and improve project separation
