+++ src/resolvers/hello-resolver.ts
@@ -1,0 +1,1 @@
+export const helloResolver = () => 'Hello World';
